Moved SetPoint to Util
