+++ src/main/java/org/usfirst/frc/team4908/robot/Util/SetPoint.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc.team4908.robot.Autonomous.Commands;
+package org.usfirst.frc.team4908.robot.Util;
 
 /**
  * @author Siggy
